First Work
Assembled basic mechanics for the game:
	-Blocking:
Placing both hands in front of the face enters blocking mode.
			-Movement:
Movement came packaged with the VR Template scene in Unity, though it still needs to be tuned to our game specifically.
			- Punching:
Controller velocity is now recorded as a variable to be used for damage calculations etc later.
Known Bugs/Issues:
-Block detection:
I detect blocks by counting how many hands are in the trigger attached to the camera (2 hands = block). For some reason the counter is starting at 1, despite being set equal to 0 in code.
-Teleport Controls:
Unity comes with VR Movement built into the template. Left joystick works fine for typical movement, but the right joystick is supposed to allow you to teleport where you are pointing the controller. Right now this is not working.