Healthbar improvements.
healthbars work for enemies ( just make a WorldCanvas on the enemy and add a healthbar prefab as a child of that).

Player healthbar is now on wrists. rotates with the hand, to make it easier to read, and can be swapped from left to right.

Healthbar now has a "hit effect" where when taking damage, it briefly displays how much damage you took. hard to explain in words but you'll know when you see it.|

Disabled Snap turnaround so you can no longer flip 180 degrees instantly.

Increased damage velocity multiplier.